requirements.txt
aiogram>=3.6.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiosqlite>=0.19.0
pydantic>=2.6.0
python-dotenv>=1.0.0

